+++ Source/KeyboardObservable.swift
@@ -20,23 +20,28 @@ public protocol KeyboardObservable: NSObjectProtocol {
 
 public extension KeyboardObservable {
     func registerForKeyboardEvents() {
-        NotificationCenter.default.addObserver(forName: UIResponder.keyboardWillShowNotification,
-                                               object: nil,
-                                               queue: .main,
-                                               using: { [weak self] notification in
-                                                   self?.keyboardWillShow(notification)
+        _ = NotificationCenter.default.addObserver(forName: UIResponder.keyboardWillShowNotification,
+                                                   object: nil,
+                                                   queue: .main,
+                                                   using: { [weak self] notification in
+                                                       self?.keyboardWillShow(notification)
         })
 
-        NotificationCenter.default.addObserver(forName: UIResponder.keyboardWillHideNotification,
-                                               object: nil,
-                                               queue: .main,
-                                               using: { [weak self] notification in
-                                                   self?.keyboardWillHide(notification)
+        _ = NotificationCenter.default.addObserver(forName: UIResponder.keyboardWillHideNotification,
+                                                   object: nil,
+                                                   queue: .main,
+                                                   using: { [weak self] notification in
+                                                       self?.keyboardWillHide(notification)
         })
     }
 
     func unregisterForKeyboardEvents() {
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillShowNotification, object: nil)
-        NotificationCenter.default.removeObserver(self, name: UIResponder.keyboardWillHideNotification, object: nil)
+        NotificationCenter.default.removeObserver(self,
+                                                  name: UIResponder.keyboardWillShowNotification,
+                                                  object: nil)
+
+        NotificationCenter.default.removeObserver(self,
+                                                  name: UIResponder.keyboardWillHideNotification,
+                                                  object: nil)
     }
 }