+++ Source/UIAlertController+Convenience.swift
@@ -12,16 +12,14 @@ public typealias AlertActionHandler = (UIAlertAction) -> Void
 /// - custom: Custom Alert Action that uses the provided title and style
 
 public enum AlertAction {
-    case okay(AlertActionHandler?)
-    case cancel(AlertActionHandler?)
-    case delete(AlertActionHandler?)
-    case done(AlertActionHandler?)
+    case okay(AlertActionHandler? = nil)
+    case cancel(AlertActionHandler? = nil)
+    case delete(AlertActionHandler? = nil)
+    case done(AlertActionHandler? = nil)
     case custom(String?, UIAlertAction.Style, AlertActionHandler?)
 }
 
-public extension AlertAction {
-    /// The corresponding UIAlertAction for each case
-
+private extension AlertAction {
     var alertAction: UIAlertAction {
         switch self {
         case let .okay(handler):
@@ -52,9 +50,29 @@ public extension UIAlertController {
     ///   - error: A value conforming to the Swift.Error protocol
     ///   - alertActions: An array of AlertActions to add as actions for the controller
 
-    convenience init(error: Error, alertActions: [AlertAction]) {
+    convenience init(error: Error,
+                     alertActions: [AlertAction]? = nil) {
         self.init(title: NSLocalizedString("Error", comment: ""),
                   message: error.localizedDescription,
+                  preferredStyle: .alert,
+                  alertActions: alertActions)
+    }
+
+    /// Convenience initializer for an UIAlertController that takes in some default
+    /// properties, as well as an array of AlertActions.
+    ///
+    /// Note: Alert actions will be added based on their ordering within the array.
+    ///
+    /// - Parameters:
+    ///   - title: Optional string for the title of the controller
+    ///   - message: Optional string for the message of the controller
+    ///   - alertActions: Optional array of AlertActions to add as actions for the controller
+
+    convenience init(title: String? = nil,
+                     message: String? = nil,
+                     alertActions: [AlertAction]? = nil) {
+        self.init(title: title,
+                  message: message,
                   preferredStyle: .alert,
                   alertActions: alertActions)
     }
@@ -73,10 +91,11 @@ public extension UIAlertController {
     convenience init(title: String? = nil,
                      message: String? = nil,
                      preferredStyle: UIAlertController.Style = .alert,
-                     alertActions: [AlertAction]) {
+                     alertActions: [AlertAction]?) {
         self.init(title: title, message: message, preferredStyle: preferredStyle)
 
-        guard alertActions.count > 0 else {
+        guard let alertActions = alertActions,
+            alertActions.count > 0 else {
             addAction(AlertAction.okay(nil).alertAction)
             return
         }