+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 1.5.0 (iOS 13 and Updates)
+
+### Updated
+
+- Recompiled the library for Xcode 11 and iOS 13
+- [KeyboardObservable](./Source/KeyboardObservable.swift)
+  - Code formatting
+- [KeyboardRespondable](./Source/KeyboardRespondable.swift)
+  - Added the ability to define multiple `ContentInsetAdjustable` views to respond to the keyboard.
+- [UIAlertController+Convenience.swift](UIAlertController+Convenience.swift)
+  - Added an additional convenience initializer
+  - Updated the nullability of some convenience initializer parameters
+- [README.md](./README.md)
+- [CHANGELOG.md](./CHANGELOG.md)
+
 ## 1.4.0 (Sub Specs)
 
 ### Added
@@ -14,6 +29,8 @@
   - Code formatting
 - [KeyboardRespondable](./Source/KeyboardRespondable.swift)
   - Code formatting
+- [README.md](./README.md)
+- [CHANGELOG.md](./CHANGELOG.md)
 
 ## 1.3.0 (Keyboard Utils)
 
@@ -44,7 +61,9 @@
 ### Updated
 
 - [UIAlertController+Convenience.swift](UIAlertController+Convenience.swift)
-- Changed the `.ok` case to `.okay` to comply with SwiftLint rule `identifier_name`
+  - Changed the `.ok` case to `.okay` to comply with SwiftLint rule `identifier_name`
+- [README.md](./README.md)
+- [CHANGELOG.md](./CHANGELOG.md)
 
 ## 1.1.1 (CHANGELOG.md)